fixed bug selection sizes fixed

+++ src/hooks/useBasket.js
@@ -1,6 +1,9 @@
-import { displayActionMessage } from '@/helpers/utils';
-import { useDispatch, useSelector } from 'react-redux';
-import { addToBasket as dispatchAddToBasket, removeFromBasket } from '@/redux/actions/basketActions';
+import { displayActionMessage } from "@/helpers/utils";
+import { useDispatch, useSelector } from "react-redux";
+import {
+  addToBasket as dispatchAddToBasket,
+  removeFromBasket,
+} from "@/redux/actions/basketActions";
 
 const useBasket = () => {
   const { basket } = useSelector((state) => ({ basket: state.basket }));
@@ -9,12 +12,14 @@ const useBasket = () => {
   const isItemOnBasket = (id) => !!basket.find((item) => item.id === id);
 
   const addToBasket = (product) => {
+    if (product.selectedSize === "")
+      return displayActionMessage("select size", "info");
     if (isItemOnBasket(product.id)) {
       dispatch(removeFromBasket(product.id));
-      displayActionMessage('Item removed from basket', 'info');
+      displayActionMessage("Item removed from cart", "info");
     } else {
       dispatch(dispatchAddToBasket(product));
-      displayActionMessage('Item added to basket', 'success');
+      displayActionMessage("Item added to cart", "success");
     }
   };
